+++ asset/files/catalog.json
@@ -22,7 +22,7 @@
             "desc": "Apple Macbook air with apple silicon",
             "price": 1099,
             "color": "#e0bfae",
-            "image": "https://i.pinimg.com/564x/d3/18/c5/d318c5256fd2be888d1542a0fd3cec54.jpg"
+            "image": "https://images.firstpost.com/wp-content/uploads/2020/11/Apple_macbook-air.jpg"
         },
         {
             "id": 4,
