+++ lib/pages/home_detail_pages.dart
@@ -1,0 +1,74 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_application_1/models/catolog.dart';
+import 'package:flutter_application_1/widgets/themes.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class HomeDetailPage extends StatelessWidget {
+  final Item catalog;
+  const HomeDetailPage({super.key, required this.catalog})
+      : assert(catalog != null);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(catalog.name),
+      ),
+      backgroundColor: MyTheme.creamColor,
+      bottomNavigationBar: Container(
+        color: Colors.white,
+        child: ButtonBar(
+          alignment: MainAxisAlignment.spaceBetween,
+          buttonPadding: EdgeInsets.zero,
+          children: [
+            "\$${catalog.price}".text.bold.xl4.red800.make(),
+            ElevatedButton(
+              onPressed: () {},
+              style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.all(
+                    MyTheme.darkBluishColor,
+                  ),
+                  shape: MaterialStateProperty.all(
+                    const StadiumBorder(),
+                  )),
+              child: "Buy".text.white.make(),
+            ).wh(100, 50)
+          ],
+        ).p32(),
+      ),
+      body: SafeArea(
+        bottom: false,
+        child: Column(
+          children: [
+            Hero(
+                    tag: Key(catalog.id.toString()),
+                    child: Image.network(catalog.image))
+                .h32(context),
+            Expanded(
+              child: VxArc(
+                  height: 30.0,
+                  arcType: VxArcType.convey,
+                  edge: VxEdge.top,
+                  child: Container(
+                    color: Colors.white,
+                    width: context.screenWidth,
+                    child: Column(
+                      children: [
+                        catalog.name.text.xl4
+                            .color(MyTheme.darkBluishColor)
+                            .bold
+                            .make(),
+                        catalog.desc.text.xl
+                            .textStyle(context.captionStyle)
+                            .make(),
+                        10.heightBox,
+                      ],
+                    ).py64(),
+                  )),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
